+++ packages/calendar/src/components/calendar-types/multi-date-calendar/UseMultiDateSelection.ts
@@ -4,15 +4,16 @@ import { OnClickDay } from "../../../types/CalendarTypes";
 import { addDayStateHighlights } from "../../../util/calendar/StateModifier";
 import { MultiDateCalendarProps } from "./MultiDateCalendar";
 import { CalendarWithMonthSwitcherProps } from "../../../features/month-switcher/CalendarWithMonthSwitcher";
-import { CalendarPanelType } from "../../../features/calendar-with-month-year-pickers/CalendarPanelType";
+import { useInternalPanelState } from "../../../features/internal-panel-state/UseInternalPanelState";
 
 export const useMultiDateSelection = <T>({
   onChange,
   value,
   statePerMonth,
+  onChangePanel,
 }: MultiDateCalendarProps<T>): CalendarWithMonthSwitcherProps<T> => {
-  const [currentPanel, setCurrentPanel] = useState<CalendarPanelType>(
-    "calendar"
+  const { currentPanel, setCurrentPanel } = useInternalPanelState(
+    onChangePanel
   );
 
   const [dateInFocus, setDateInFocus] = useState(() => new Date());