+++ packages/grid/src/features/standard-table/components/StandardTableRowList.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { useMemo, useRef } from "react";
+import { useEffect, useMemo, useRef } from "react";
 import { multitypeComparator } from "../features/sorting/MultitypeComparator";
 import { useColumnValueResolver } from "../hooks/UseColumnValueResolver";
 import {
@@ -8,6 +8,8 @@ import {
 } from "../hooks/UseStandardTableConfig";
 import { StandardTableVariant } from "./StandardTable";
 import { StandardTableRow } from "./StandardTableRow";
+import { SummaryRowSwitcher } from "../features/summary-row/components/SummaryRowSwitcher";
+import { filterItemsOnEnabledCheckboxes } from "../util/FilterItemsOnEnabledCheckboxes";
 
 interface StandardTableContentProps<TItem> {
   items?: Array<TItem>;
@@ -29,7 +31,14 @@ export const StandardTableRowList = React.memo(function StandardTableRowList<
    * rows after sorting.
    */
   const sortCounterRef = useRef(0);
-  const { keyResolver, disableInfiniteList } = useStandardTableConfig();
+
+  const shiftPressedRef = useRef(false);
+
+  const {
+    keyResolver,
+    disableInfiniteList,
+    checkboxDisabledResolver,
+  } = useStandardTableConfig();
   const {
     sortOrder: { sortBy, desc },
   } = useStandardTableState();
@@ -55,18 +64,50 @@ export const StandardTableRowList = React.memo(function StandardTableRowList<
     return sortedList;
   }, [items, valueResolver, desc]);
 
+  const idListForEnabledItems = useMemo(
+    () =>
+      sortedItems
+        .filter(filterItemsOnEnabledCheckboxes(checkboxDisabledResolver))
+        .map((l) => keyResolver(l)),
+    [sortedItems, checkboxDisabledResolver, keyResolver]
+  );
+
+  useEffect(() => {
+    const keyUp = (ev: KeyboardEvent) => {
+      if (ev.key === "Shift") {
+        shiftPressedRef.current = false;
+      }
+    };
+
+    const keyDown = (ev: KeyboardEvent) => {
+      if (ev.key === "Shift") {
+        shiftPressedRef.current = true;
+      }
+    };
+
+    document.addEventListener("keyup", keyUp);
+    document.addEventListener("keydown", keyDown);
+    return () => {
+      document.removeEventListener("keyup", keyUp);
+      document.removeEventListener("keydown", keyDown);
+    };
+  }, []);
+
   return (
     <React.Fragment key={sortCounterRef.current}>
       {sortedItems.map((item, index) => (
         <StandardTableRow
           alwaysVisible={disableInfiniteList || sortedItems.length < 30}
           item={item}
+          idListForEnabledItems={idListForEnabledItems}
           key={keyResolver(item)}
           colIndexOffset={colIndexOffset}
           rowIndex={index + rowIndexOffset}
           numRows={sortedItems.length}
+          shiftPressedRef={shiftPressedRef}
         />
       ))}
+      <SummaryRowSwitcher items={sortedItems} />
     </React.Fragment>
   );
 });