+++ packages/tooltip/src/index.ts
@@ -2,3 +2,5 @@ export * from "./components/popover/ActionPrompt";
 export * from "./components/popover/Popover";
 export * from "./components/button-with-popover/ButtonWithPopover";
 export * from "./components/tooltip/Tooltip";
+export * from "./hooks/UseTippyInstance";
+export * from "./types";